Support building the fetcher with -Wimplicit-atomic-properties
This is mostly just adding atomic explicitly to declarations. This CL also
removes some unneeded @synthesize directives and provides explicit
implementations of chunkSize and delegateCallbackQueue.

The meaning of "atomic" for accessors explicitly implemented is kinda vague,
so some of these might be arguably better described as nonatomic, but this
CL isn't trying to resolve that conundrum.