Safely handle when the NSURLSession has been swizzled and its delegate replaced with a proxy. (#142)
Safely handle the NSURLSession's delegate being a proxy.

Several third-party libraries are known to swizzle NSURLSession and replace the delegate with a proxy, which caused the GTMSessionFetcherService to misbehave and think the fetcher had a delegate dispatcher when it didn't. By replace a pointer equality check with an isKindOfClass: check, this can be avoided.