Fix background task behavior with background sessions. (#186)
Fix background task behavior with background sessions.

Previously was checking `useBackgroundSession` for the active state, while that property is documented only as the client's expressed preference that the fetch use a background session. Even with `useBackgroundSession = NO`, if a `sessionIdentifer` is set before beginning the fetch than a background session will be used (and `usingBackgroundSession` will be `YES`.

Because `useBackgroundSession` was being checked instead, a background task would be created but shouldn't have been.

Also fixed a potential issue if a fetch for a data: URL is created and `useBackgroundSession == YES` or a session identifier was provided that could lead to attempting a background session for a data: URL fetch.