Revert "Work around early iOS 13 betas still in use."
This reverts commit 67d6e1ba577b4fe902d3ce6fd4374bf841e9b902.

Original PR was intended to be a short-term change due to a large-ish number of users still remaining on early iOS 13 public betas. Had planned to reevaluate that decision eventually, and have come to the conclusion we should no longer maintain the fix rather than retain the standard mechanism for checking for API support.